规格参数 增 删 改 查

+++ mingrui-shop-service/mingrui-shop-service-xxx/src/main/java/com/baidu/shop/serviceImpl/SpecificationServiceImpl.java
@@ -4,10 +4,14 @@ import com.alibaba.fastjson.JSONObject;
 import com.baidu.shop.base.BaseApiService;
 import com.baidu.shop.base.Result;
 import com.baidu.shop.dto.SpecGroupDTO;
+import com.baidu.shop.dto.SpecParamDTO;
 import com.baidu.shop.entity.SpecGroupEntity;
+import com.baidu.shop.entity.SpecParamEntity;
 import com.baidu.shop.mapper.SpecGroupMapper;
+import com.baidu.shop.mapper.SpecParamMapper;
 import com.baidu.shop.service.SpecificationService;
 import com.baidu.shop.utils.BaiduBeanUtil;
+import com.baidu.shop.utils.ObjectUtil;
 import org.springframework.transaction.annotation.Transactional;
 import tk.mybatis.mapper.entity.Example;
 
@@ -25,6 +29,9 @@ public class SpecificationServiceImpl extends BaseApiService implements Specific
 
     @Resource
     private SpecGroupMapper specificationMapper;
+
+    @Resource
+    private SpecParamMapper specParamMapper;
 
     @Override
     public Result<List<SpecGroupEntity>> list(SpecGroupDTO specGroupDTO) {
@@ -54,7 +61,54 @@ public class SpecificationServiceImpl extends BaseApiService implements Specific
     @Override
     public Result<JSONObject> delete(Integer id) {
 
+        //删除规格组之前需要先判断一下当前规格组下是否有规格参数
+        Example example = new Example(SpecParamEntity.class);
+        example.createCriteria().andEqualTo("groupId",id);
+
+        List<SpecParamEntity> specParamEntities = specParamMapper.selectByExample(example);
+        if (specParamEntities.size() >= 1) return this.setResultError("该规格组下有规格参数,无法删除");
         specificationMapper.deleteByPrimaryKey(id);
+
+        return this.setResultSuccess();
+    }
+
+    @Override
+    public Result<List<SpecParamEntity>> list(SpecParamDTO specParamDTO) {
+
+        Example example = new Example(SpecParamEntity.class);
+        Example.Criteria criteria = example.createCriteria();
+
+        if (ObjectUtil.isNotNull(specParamDTO.getGroupId()))
+            criteria.andEqualTo("groupId",specParamDTO.getGroupId());
+
+        if (ObjectUtil.isNotNull(specParamDTO.getCid()))
+            criteria.andEqualTo("cid",specParamDTO.getCid());
+
+        List<SpecParamEntity> specParamEntities = specParamMapper.selectByExample(example);
+
+        return this.setResultSuccess(specParamEntities);
+    }
+
+    @Override
+    public Result<JSONObject> saveSpecParam(SpecParamDTO specParamDTO) {
+
+        specParamMapper.insertSelective(BaiduBeanUtil.copyProperties(specParamDTO,SpecParamEntity.class));
+
+        return this.setResultSuccess();
+    }
+
+    @Override
+    public Result<JSONObject> editSpecParam(SpecParamDTO specParamDTO) {
+
+        specParamMapper.updateByPrimaryKeySelective(BaiduBeanUtil.copyProperties(specParamDTO,SpecParamEntity.class));
+
+        return this.setResultSuccess();
+    }
+
+    @Override
+    public Result<JSONObject> deleteSpecParamById(Integer id) {
+
+        specParamMapper.deleteByPrimaryKey(id);
 
         return this.setResultSuccess();
     }
